+++ database/schema/07-pricefeed.sql
@@ -24,7 +24,7 @@ CREATE TABLE token_price
 
     unit_name  TEXT                        NOT NULL REFERENCES token_unit (denom) UNIQUE,
     price      DECIMAL                     NOT NULL,
-    market_cap BIGINT                      NOT NULL,
+    market_cap NUMERIC                      NOT NULL,
     timestamp  TIMESTAMP WITHOUT TIME ZONE NOT NULL
 );
 CREATE INDEX token_price_timestamp_index ON token_price (timestamp);
@@ -51,7 +51,7 @@ CREATE TABLE token_price_history
     id         SERIAL                      NOT NULL PRIMARY KEY,
     unit_name  TEXT                        NOT NULL REFERENCES token_unit (denom),
     price      DECIMAL                     NOT NULL,
-    market_cap BIGINT                      NOT NULL,
+    market_cap NUMERIC                      NOT NULL,
     timestamp  TIMESTAMP WITHOUT TIME ZONE NOT NULL,
     CONSTRAINT unique_price_for_timestamp UNIQUE (unit_name, timestamp)
 );
